Add --sync --sync-fast as convenient aliases
Also avoid redundant delete when updating files.
Also fixed dummy mode on file/dir type mismatch
(always assume destinations are not existing in dummyMode. This is safe
because destinations are always deleted (or printed as deleted in dummyMode)
beforehand.
Fix verbose messages when updating mtime (but --preserve is not yet implemented
so this does nothing).